Avoid double conversion warnings

+++ stb_vorbis.c
@@ -1253,7 +1253,7 @@ static void compute_window(int n, float *window)
 {
    int n2 = n >> 1, i;
    for (i=0; i < n2; ++i)
-      window[i] = (float) sin(0.5 * M_PI * square((float) sin((i - 0 + 0.5) / n2 * 0.5 * M_PI)));
+      window[i] = (float) sinf(0.5f * M_PI * square((float) sin((i - 0 + 0.5f) / n2 * 0.5f * M_PI)));
 }
 
 static void compute_bitreverse(int n, uint16 *rev)
